Update README.md
This repository is part of my learning roadmap in cybersecurity. It contains notes, walkthroughs, and commands that help me practice and document what I’m learning in TryHackMe, Hack The Box, and other beginner-friendly platforms. I’ll keep updating it regularly with new insights and improvements as I grow in this field.